Certificate tests

+++ src/Utils/libs/cron/CertificateCron.js
@@ -1,21 +1,31 @@
 import cron from "node-cron";
-import TreeModel from "../../Models/TreeModel.js";
-import CertificateModel from "../../Models/CertificateModel.js";
+import TreeModel from "../../../Models/TreeModel.js";
+import CertificateModel from "../../../Models/CertificateModel.js";
 
-cron.schedule("0 0 * * *", async () => {
-  try {
-    const expiredCertificates = await CertificateModel.find({ expiresAt: { $lt: new Date() } });
-
-    for (const cert of expiredCertificates) {
-      const tree = await TreeModel.findById(cert.id_tree);
-      if (tree) {
-        tree.available_quantity += cert.quantity;
-        await tree.save();
+export const startCertificateExpirationJob = () => {
+  cron.schedule("0 0 * * *", async () => {
+    try {
+      
+      const currentDate = new Date();
+      const expiredCertificates = await CertificateModel.find({ finalDate: { $lt: currentDate } });
+      if (expiredCertificates.length === 0) {
+        return;
       }
-
-      await CertificateModel.findByIdAndDelete(cert._id);
+      await Promise.all(
+        expiredCertificates.map(async (cert) => {
+         
+           await TreeModel.updateOne(
+            { _id: cert.id_tree },
+            {
+              $inc: { available_quantity: cert.quantity },
+            }
+          );
+        
+          await CertificateModel.findByIdAndDelete(cert._id);
+        })
+      );
+    } catch (error) {
+      console.log(error);
     }
-  } catch (error) {
-    console.log(error);
-  }
-});
+  });
+};
